requirements.txt
torch
transformers
pillow
pickleshare
numpy
tensorflow-cpu
Flask
flask_cors
gevent